Change the model and entity structure
The result is returning as a json format like {'result': res}
In the model fromJson factory function you are trying to get the value as
     addResult:  json['addResult'],
      subResult: json['subResult'],
      mulResult: json['mulResult'],
      divResult: json['divResult'],
Those are not there in the json.
And result is a double value regardless of addition or subtraction or others.
I have fixed. Please look into it.